Persist search query in firestore
