+++ android/app/src/main/kotlin/com/gchristov/newsfeed/android/app/DependencyInjector.kt
@@ -2,6 +2,7 @@ package com.gchristov.newsfeed.android.app
 
 import android.content.Context
 import com.gchristov.newsfeed.android.common.navigation.NavigationModule
+import com.gchristov.newsfeed.multiplatform.auth.data.MplAuthDataModule
 import com.gchristov.newsfeed.multiplatform.common.firebase.MplCommonFirebaseModule
 import com.gchristov.newsfeed.multiplatform.common.kotlin.AppContext
 import com.gchristov.newsfeed.multiplatform.common.kotlin.MplCommonKotlinModule
@@ -24,6 +25,7 @@ internal fun DependencyInjector.initialise(context: Context) {
     AppContext = context
     val modules = mutableListOf<DependencyModule>().apply {
         // Multiplatform modules
+        add(MplAuthDataModule)
         add(MplCommonKotlinModule)
         add(MplCommonNetworkModule)
         add(MplCommonPersistenceModule)
