+++ multiplatform/feed/data/src/commonMain/kotlin/com/gchristov/newsfeed/multiplatform/feed/data/api/ApiFeedResponse.kt
@@ -1,29 +1,30 @@
 package com.gchristov.newsfeed.multiplatform.feed.data.api
 
+import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class ApiFeedResponse(
-    val response: ApiFeed,
+    @SerialName("response") val response: ApiFeed,
 ) {
     @Serializable
     data class ApiFeed(
-        val pages: Int,
-        val currentPage: Int,
-        val results: List<ApiFeedItem>
+        @SerialName("pages") val pages: Int,
+        @SerialName("currentPage") val currentPage: Int,
+        @SerialName("results") val results: List<ApiFeedItem>
     )
 
     @Serializable
     data class ApiFeedItem(
-        val id: String,
-        val apiUrl: String,
-        val webPublicationDate: String,
-        val fields: ApiFeedItemFields? = null,
+        @SerialName("id") val id: String,
+        @SerialName("apiUrl") val apiUrl: String,
+        @SerialName("webPublicationDate") val webPublicationDate: String,
+        @SerialName("fields") val fields: ApiFeedItemFields? = null,
     )
 
     @Serializable
     data class ApiFeedItemFields(
-        val headline: String? = null,
-        val thumbnail: String? = null,
+        @SerialName("headline") val headline: String? = null,
+        @SerialName("thumbnail") val thumbnail: String? = null,
     )
 }
