+++ multiplatform/feed/data/src/commonMain/kotlin/com/gchristov/newsfeed/multiplatform/feed/data/db/DbFeedFilter.kt
@@ -1,0 +1,9 @@
+package com.gchristov.newsfeed.multiplatform.feed.data.db
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+internal data class DbFeedFilter(
+    @SerialName("query") val query: String,
+)
