+++ multiplatform/feed/data/src/commonMain/kotlin/com/gchristov/newsfeed/multiplatform/feed/data/model/FeedFilter.kt
@@ -1,0 +1,5 @@
+package com.gchristov.newsfeed.multiplatform.feed.data.model
+
+data class FeedFilter(
+    val query: String
+)
